feat: add basics for power gen model

+++ src/math/moments.rs
@@ -1,5 +1,5 @@
 pub fn mean(input: Vec<f64>) -> f64 {
-    let len = input.len();
+    let len = input.len() as f64;
 
     let mut result = 0f64;
     for e in input {
@@ -15,7 +15,7 @@ pub fn standard_deviation(input: Vec<f64>) -> f64 {
     let mut variance: f64 = 0f64;
 
     for e in input {
-        variance += (e-mean).pow(2);
+        variance += (e-mean).powf(2.0);
     }
 
     return variance.sqrt();
